More cleanup
git-svn-id: https://svn.smr.arpa.emr.it/radar/trunk@1937 5e904517-b230-484e-af0d-ad602924471e

+++ include/volume.h
@@ -74,7 +74,7 @@ public:
      * Riempie un array di float con i dati del raggio convertiti in DB
      * Se l'array è più lungo del raggio, setta gli elementi extra a missing.
      */
-    void read_beam_db(unsigned az, float* out, unsigned out_size, float missing=0) const
+    void read_beam_db(unsigned az, double* out, unsigned out_size, float missing=0) const
     {
         using namespace std;
